rename src/run_analysis.py as src/analysis.py
